+++ .storybook/preview.ts
@@ -30,6 +30,9 @@ const preview: Preview = {
   },
   parameters: {
     layout: "fullscreen",
+    a11y: {
+      test: "error",
+    },
     controls: {
       expanded: true,
       sort: "requiredFirst",